requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Authentication & Security
PyJWT>=2.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.2

# Real-time & WebSockets
websockets==12.0
python-socketio==5.10.0
aiohttp==3.9.1

# API Clients
httpx==0.25.1
requests==2.31.0

# Geolocation & Maps
geopy==2.4.1

# Media Processing
Pillow==10.1.0
opencv-python==4.8.1.78
moviepy==1.0.3

# Utilities
pydantic==2.5.0
pydantic-settings==2.1.0
python-dateutil==2.8.2
pytz==2023.3

# Logging
colorlog==6.8.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1
